+++ tests/testthat/test-table_heat.R
@@ -1,0 +1,7 @@
+context("Checking table_heat")
+
+test_that("table_heat ...",{
+
+
+})
+